refactor: refactor LocaleSwitcher
rename tool-bar to toolbar
create locale-tool
edit messages
