--- src/components/tool-bar/locale-tool.tsx
@@ -1,2 +1,0 @@
-import LocaleSwitcher from '../translation/LocaleSwitcher'
-export { LocaleSwitcher as LocaleTool }
