--- src/components/translation/LocaleSwitcher.tsx
@@ -1,24 +1,0 @@
-import { useLocale, useTranslations } from 'next-intl'
-import LocaleSwitcherSelect from './LocaleSwitcherSelect'
-
-export default function LocaleSwitcher() {
-  const t = useTranslations('LocaleSwitcher')
-  const locale = useLocale()
-
-  return (
-    <LocaleSwitcherSelect
-      defaultValue={locale}
-      items={[
-        {
-          value: 'en',
-          label: t('en'),
-        },
-        {
-          value: 'de',
-          label: t('de'),
-        },
-      ]}
-      label={t('label')}
-    />
-  )
-}
